Delete Profile(READ DESC)
Hi!
So, this doesnt technically work, it rewrites the entire Profile.txt as empty, which isn't right, I need to go thru the saveProfile function and copy some more code from there to write properly - feel free to have a go at fixing it, since im prob not gonna have another chance to fix it tonight.

Also, when we're finishing the code, we'll have to load the text file with Blank Profiles otherwise the program crashes (may need a seperate function to check if file is empty, and if it is then initialise it with 4 blank profiles)